2007-01-02  Pace Willisson  <pace@rhino.pacew.com>
	* V21

	* enable mission failures (and make command line arg to force all
	to succeed again); allow jumbled arg order in draw/clear rectangle;
	give symbolic names to xMODE flags (for cloud testing); failure
	animation; data type size problem reading NTABLE.DAT;

+++ intro.c
@@ -165,7 +165,7 @@ struct intro_image intro_images[INTRO_IMAGE_COUNT];
 int
 read_img_frame (FILE *inf, struct intro_image *ip)
 {
-	int len;
+	long len;
 	char compressed[64 * 1024];
 	int ilen;
 	int idx;
